+++ lib/src/ilist/ilist.dart
@@ -6,7 +6,6 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:fast_immutable_collections/src/base/hash.dart";
 import "package:meta/meta.dart";
 
-
 import "ilist_of_2.dart";
 import "l_add.dart";
 import "l_add_all.dart";
@@ -34,6 +33,7 @@ class IList<T> // ignore: must_be_immutable
 
   /// Create an [IList] from any [Iterable] and a [ConfigList].
   /// Fast, if the Iterable is another [IList].
+  /// If [iterable] is null, return an empty [IList].
   factory IList.withConfig(
     Iterable<T>? iterable,
     ConfigList config,
@@ -83,7 +83,8 @@ class IList<T> // ignore: must_be_immutable
     required ConfigList? config,
   }) {
     List<T> list = iset.toList(growable: false, compare: compare);
-    return IList._unsafe(LFlat<T>.unsafe(list), config: config ?? defaultConfig);
+    return IList._unsafe(LFlat<T>.unsafe(list),
+        config: config ?? defaultConfig);
   }
 
   /// **Unsafe constructor. Use this at your own peril.**
@@ -96,10 +97,43 @@ class IList<T> // ignore: must_be_immutable
   /// Note you can optionally disallow unsafe constructors ([ImmutableCollection]) in the global
   /// configuration by doing: `ImmutableCollection.disallowUnsafeConstructors = true` (and then
   /// optionally preventing further configuration changes by calling `lockConfig()`).
-  IList.unsafe(List<T> list, {required this.config}) : _l = LFlat<T>.unsafe(list) {
+  IList.unsafe(List<T> list, {required this.config})
+      : _l = LFlat<T>.unsafe(list) {
     if (ImmutableCollection.disallowUnsafeConstructors)
       throw UnsupportedError("IList.unsafe is disallowed.");
   }
+
+  /// If [iterable] is null, return null.
+  ///
+  /// Otherwise, Create an [IList] from the [Iterable].
+  /// Fast, if the Iterable is another [IList].
+  ///
+  /// This static factory is useful for implementing a `copyWith` method
+  /// that accept an [Iterable]. For example:
+  ///
+  /// ```
+  ///  IList<String> names;
+  ///
+  ///  Students copyWith({ Iterable<String>? names }) =>
+  ///    Students(names: IList.orNull(names) ?? this.names);
+  /// ```
+  ///
+  /// Of course, if your `copyWith` accepts an [IList], this is not necessary:
+  ///
+  /// ```
+  ///  IList<String> names;
+  ///
+  ///  Students copyWith({ IList<String>? names }) =>
+  ///    Students(names: names ?? this.names);
+  /// ```
+  ///
+  static IList<T>? orNull<T>(
+    Iterable<T>? iterable, [
+    ConfigList? config,
+  ]) =>
+      (iterable == null)
+          ? null
+          : IList.withConfig(iterable, config ?? defaultConfig);
 
   /// Returns an empty [IList], with the given configuration. If a
   /// configuration is not provided, it will use the default configuration.
@@ -114,7 +148,8 @@ class IList<T> // ignore: must_be_immutable
   /// [ImmutableCollection.isConfigLocked],[flushFactor], [defaultConfig]
   static void resetAllConfigurations() {
     if (ImmutableCollection.isConfigLocked)
-      throw StateError("Can't change the configuration  of immutable collections.");
+      throw StateError(
+          "Can't change the configuration  of immutable collections.");
     IList.flushFactor = _defaultFlushFactor;
     IList.defaultConfig = _defaultConfig;
   }
@@ -139,7 +174,8 @@ class IList<T> // ignore: must_be_immutable
   static set defaultConfig(ConfigList config) {
     if (_defaultConfig == config) return;
     if (ImmutableCollection.isConfigLocked)
-      throw StateError("Can't change the configuration of immutable collections.");
+      throw StateError(
+          "Can't change the configuration of immutable collections.");
     _defaultConfig = config;
   }
 
@@ -147,7 +183,8 @@ class IList<T> // ignore: must_be_immutable
   static set flushFactor(int value) {
     if (_flushFactor == value) return;
     if (ImmutableCollection.isConfigLocked)
-      throw StateError("Can't change the configuration of immutable collections.");
+      throw StateError(
+          "Can't change the configuration of immutable collections.");
     if (value > 0)
       _flushFactor = value;
     else
@@ -158,7 +195,8 @@ class IList<T> // ignore: must_be_immutable
   static set asyncAutoflush(bool value) {
     if (_asyncAutoflush == value) return;
     if (ImmutableCollection.isConfigLocked)
-      throw StateError("Can't change the configuration of immutable collections.");
+      throw StateError(
+          "Can't change the configuration of immutable collections.");
     _asyncAutoflush = value;
   }
 
@@ -236,15 +274,18 @@ class IList<T> // ignore: must_be_immutable
   IList._unsafe(this._l, {required this.config});
 
   /// **Unsafe**.
-  IList._unsafeFromList(List<T> list, {required this.config}) : _l = LFlat<T>.unsafe(list);
+  IList._unsafeFromList(List<T> list, {required this.config})
+      : _l = LFlat<T>.unsafe(list);
 
   /// Creates a list with `identityEquals` (compares the internals by `identity`).
-  IList<T> get withIdentityEquals =>
-      config.isDeepEquals ? IList._unsafe(_l, config: config.copyWith(isDeepEquals: false)) : this;
+  IList<T> get withIdentityEquals => config.isDeepEquals
+      ? IList._unsafe(_l, config: config.copyWith(isDeepEquals: false))
+      : this;
 
   /// Creates a list with `deepEquals` (compares all list items by equality).
-  IList<T> get withDeepEquals =>
-      config.isDeepEquals ? this : IList._unsafe(_l, config: config.copyWith(isDeepEquals: true));
+  IList<T> get withDeepEquals => config.isDeepEquals
+      ? this
+      : IList._unsafe(_l, config: config.copyWith(isDeepEquals: true));
 
   /// See also: [ConfigList]
   bool get isDeepEquals => config.isDeepEquals;
@@ -325,9 +366,11 @@ class IList<T> // ignore: must_be_immutable
     }
 
     if (other is List<T>)
-      return const ListEquality<dynamic>().equals(UnmodifiableListFromIList<T>(this), other);
+      return const ListEquality<dynamic>()
+          .equals(UnmodifiableListFromIList<T>(this), other);
 
-    if (other is HashSet) throw StateError("Can't compare to HashSet (which is unordered).");
+    if (other is HashSet)
+      throw StateError("Can't compare to HashSet (which is unordered).");
 
     return const IterableEquality<dynamic>().equals(_l, other);
   }
@@ -338,7 +381,8 @@ class IList<T> // ignore: must_be_immutable
   /// one by one.
   bool unorderedEqualItems(covariant Iterable? other) {
     if (other == null) return false;
-    if (identical(this, other) || (other is IList<T> && same(other))) return true;
+    if (identical(this, other) || (other is IList<T> && same(other)))
+      return true;
     return const UnorderedIterableEquality<dynamic>().equals(_l, other);
   }
 
@@ -353,7 +397,8 @@ class IList<T> // ignore: must_be_immutable
     if (_isUnequalByHashCode(other)) return false;
 
     return config == other!.config &&
-        (identical(_l, other._l) || (flush._l as LFlat).deepListEquals(other.flush._l as LFlat));
+        (identical(_l, other._l) ||
+            (flush._l as LFlat).deepListEquals(other.flush._l as LFlat));
   }
 
   /// Return `true` if other is `null` or the cached [hashCodes] proves the
@@ -365,7 +410,9 @@ class IList<T> // ignore: must_be_immutable
   /// means we don't have this information yet, and we don't calculate it.
   bool _isUnequalByHashCode(IList? other) {
     return (other == null) ||
-        (_hashCode != null && other._hashCode != null && _hashCode != other._hashCode);
+        (_hashCode != null &&
+            other._hashCode != null &&
+            _hashCode != other._hashCode);
   }
 
   /// Will return `true` only if the lists internals are the same instances
@@ -437,7 +484,8 @@ class IList<T> // ignore: must_be_immutable
     // If the outer list is used, it will be flushed before the source lists.
     // If the source lists are not used directly, they will not flush
     // unnecessarily, and also may be garbage collected.
-    result._counter = max(_counter, ((items is IList<T>) ? (items)._counter : 0)) + 1;
+    result._counter =
+        max(_counter, ((items is IList<T>) ? (items)._counter : 0)) + 1;
 
     return result;
   }
@@ -472,7 +520,9 @@ class IList<T> // ignore: must_be_immutable
   ///
   IList<T> remove(T item) {
     final L<T> result = _l.remove(item);
-    return identical(result, _l) ? this : IList<T>._unsafe(result, config: config);
+    return identical(result, _l)
+        ? this
+        : IList<T>._unsafe(result, config: config);
   }
 
   /// Removes all occurrences of all [items] from this list.
@@ -482,7 +532,9 @@ class IList<T> // ignore: must_be_immutable
   ///
   IList<T> removeAll(Iterable<T?> items) {
     final L<T> result = _l.removeAll(items);
-    return identical(result, _l) ? this : IList<T>._unsafe(result, config: config);
+    return identical(result, _l)
+        ? this
+        : IList<T>._unsafe(result, config: config);
   }
 
   /// Removes all occurrences of [item] from this list.
@@ -497,7 +549,9 @@ class IList<T> // ignore: must_be_immutable
   ///
   IList<T> removeMany(T item) {
     final L<T> result = _l.removeMany(item);
-    return identical(result, _l) ? this : IList<T>._unsafe(result, config: config);
+    return identical(result, _l)
+        ? this
+        : IList<T>._unsafe(result, config: config);
   }
 
   /// Removes all nulls from this list.
@@ -520,7 +574,8 @@ class IList<T> // ignore: must_be_immutable
 
   /// Removes the first instance of the element, if it exists in the list.
   /// Otherwise, adds it to the list.
-  IList<T> toggle(T element) => contains(element) ? remove(element) : add(element);
+  IList<T> toggle(T element) =>
+      contains(element) ? remove(element) : add(element);
 
   /// Returns the object at the given [index] in the list or throws a [RangeError] if [index] is out
   /// of bounds.
@@ -803,7 +858,8 @@ class IList<T> // ignore: must_be_immutable
   /// Note: Not very efficient at the moment (will be improved in the future).
   /// Please use for a small number of items.
   ///
-  IList<T> sortLike(Iterable<T> ordering) => IList._unsafe(_l.sortLike(ordering), config: config);
+  IList<T> sortLike(Iterable<T> ordering) =>
+      IList._unsafe(_l.sortLike(ordering), config: config);
 
   /// Divides the list into two.
   /// The first one contains all items which satisfy the provided [test].
@@ -957,7 +1013,8 @@ class IList<T> // ignore: must_be_immutable
   /// return the unchanged list if no item satisfies the [test].
   /// - If [addIfNotFound] is `true`, add the item to the end of the list
   /// if no item satisfies the [test].
-  IList<T> replaceFirstWhere(bool Function(T item) test, T to, {bool addIfNotFound = false}) {
+  IList<T> replaceFirstWhere(bool Function(T item) test, T to,
+      {bool addIfNotFound = false}) {
     var index = indexWhere(test);
     return (index != -1)
         ? put(index, to)
@@ -1081,8 +1138,10 @@ class IList<T> // ignore: must_be_immutable
     _count();
     var _length = length;
     start ??= _length;
-    if (start < 0) throw ArgumentError.value(start, "index", "Index out of range");
-    for (int i = min(start, _length - 1); i >= 0; i--) if (this[i] == element) return i;
+    if (start < 0)
+      throw ArgumentError.value(start, "index", "Index out of range");
+    for (int i = min(start, _length - 1); i >= 0; i--)
+      if (this[i] == element) return i;
     return -1;
   }
 
@@ -1108,8 +1167,10 @@ class IList<T> // ignore: must_be_immutable
     _count();
     var _length = length;
     start ??= _length;
-    if (start < 0) throw ArgumentError.value(start, "index", "Index out of range");
-    for (int i = min(start, _length - 1); i >= 0; i--) if (test(this[i])) return i;
+    if (start < 0)
+      throw ArgumentError.value(start, "index", "Index out of range");
+    for (int i = min(start, _length - 1); i >= 0; i--)
+      if (test(this[i])) return i;
     return -1;
   }
 
@@ -1132,7 +1193,8 @@ class IList<T> // ignore: must_be_immutable
   ///
   IList<T> replaceRange(int start, int end, Iterable<T> replacement) {
     // TODO: Still need to implement efficiently.
-    return IList._unsafeFromList(toList(growable: true)..replaceRange(start, end, replacement),
+    return IList._unsafeFromList(
+        toList(growable: true)..replaceRange(start, end, replacement),
         config: config);
   }
 
@@ -1165,7 +1227,8 @@ class IList<T> // ignore: must_be_immutable
   ///
   IList<T> fillRange(int start, int end, [T? fillValue]) {
     // TODO: Still need to implement efficiently.
-    return IList._unsafeFromList(toList(growable: false)..fillRange(start, end, fillValue),
+    return IList._unsafeFromList(
+        toList(growable: false)..fillRange(start, end, fillValue),
         config: config);
   }
 
@@ -1216,7 +1279,8 @@ class IList<T> // ignore: must_be_immutable
   /// If `end` is equal to `start`, then the returned list is empty.
   IList<T> sublist(int start, [int? end]) {
     // TODO: Still need to implement efficiently.
-    return IList._unsafeFromList(toList(growable: false).sublist(start, end), config: config);
+    return IList._unsafeFromList(toList(growable: false).sublist(start, end),
+        config: config);
   }
 
   /// The [replace] method is the equivalent of `operator []=` for the [IList].
@@ -1236,7 +1300,8 @@ class IList<T> // ignore: must_be_immutable
   /// The [index] value must be non-negative and no greater than [length].
   IList<T> insert(int index, T element) {
     // TODO: Still need to implement efficiently.
-    return IList._unsafeFromList(toList(growable: true)..insert(index, element), config: config);
+    return IList._unsafeFromList(toList(growable: true)..insert(index, element),
+        config: config);
   }
 
   /// Inserts all objects of [iterable] at position [index] in this list.
@@ -1248,7 +1313,8 @@ class IList<T> // ignore: must_be_immutable
   /// The [index] value must be non-negative and no greater than [length].
   IList<T> insertAll(int index, Iterable<T> iterable) {
     // TODO: Still need to implement efficiently.
-    return IList._unsafeFromList(toList(growable: true)..insertAll(index, iterable),
+    return IList._unsafeFromList(
+        toList(growable: true)..insertAll(index, iterable),
         config: config);
   }
 
@@ -1276,7 +1342,8 @@ class IList<T> // ignore: must_be_immutable
   ///
   /// If you want to recover the removed item, you can pass a mutable [removedItem].
   ///
-  IList<T> removeLast([Output<T>? removedItem]) => removeAt(length - 1, removedItem);
+  IList<T> removeLast([Output<T>? removedItem]) =>
+      removeAt(length - 1, removedItem);
 
   /// Removes the objects in the range [start] inclusive to [end] exclusive.
   ///
@@ -1380,7 +1447,8 @@ class IList<T> // ignore: must_be_immutable
   /// If [iterable] depends on this list in some other way, no guarantees are
   /// made.
   ///
-  IList<T> setRange(int start, int end, Iterable<T> iterable, [int skipCount = 0]) {
+  IList<T> setRange(int start, int end, Iterable<T> iterable,
+      [int skipCount = 0]) {
     // TODO: Still need to implement efficiently.
     var list = toList(growable: true);
     list.setRange(start, end, iterable, skipCount);
@@ -1438,7 +1506,8 @@ abstract class L<T> implements Iterable<T> {
 
   // TODO: Still need to implement efficiently.
   /// Removes the first occurrence of [element] from this list.
-  L<T> remove(T element) => !contains(element) ? this : LFlat<T>.unsafe(unlock..remove(element));
+  L<T> remove(T element) =>
+      !contains(element) ? this : LFlat<T>.unsafe(unlock..remove(element));
 
   L<T> removeAll(Iterable<T?> elements) {
     var list = unlock;
@@ -1450,8 +1519,9 @@ abstract class L<T> implements Iterable<T> {
   }
 
   // TODO: Still need to implement efficiently.
-  L<T> removeMany(T element) =>
-      !contains(element) ? this : LFlat<T>.unsafe(unlock..removeWhere((e) => e == element));
+  L<T> removeMany(T element) => !contains(element)
+      ? this
+      : LFlat<T>.unsafe(unlock..removeWhere((e) => e == element));
 
   // TODO: Still need to implement efficiently.
   /// If the list has more than `maxLength` elements, removes the last elements so it remains
@@ -1492,7 +1562,8 @@ abstract class L<T> implements Iterable<T> {
     Set<T> newSet = Set.of(this);
     Set<T> intersection = orderingSet.intersection(newSet);
     Set<T> difference = newSet.difference(orderingSet);
-    List<T> result = ordering.where((element) => intersection.contains(element)).toList();
+    List<T> result =
+        ordering.where((element) => intersection.contains(element)).toList();
     result.addAll(difference);
     return LFlat<T>.unsafe(result);
   }
@@ -1618,3 +1689,4 @@ class InternalsForTestingPurposesIList {
 }
 
 // /////////////////////////////////////////////////////////////////////////////
+